+++ src/main/java/com/tjalia/userprofile/interactor/DeleteUserProfile.java
@@ -1,0 +1,8 @@
+package com.tjalia.userprofile.interactor;
+
+import com.tjalia.userprofile.dto.response.base.MessageResponse;
+
+public interface DeleteUserProfile {
+
+    MessageResponse execute(Long id);
+}